- add 스크롤 초기화 x

+++ data/repository/src/main/java/com/lago/repository/mapper/ResponseToModelMapper.kt
@@ -67,7 +67,7 @@ internal fun MovieDetailResponse.toMovieDetail() = MovieDetail(
     tagline,
     title,
     video,
-    voteAverage,
+    voteAverage.toFloat(),
     voteCount
 )
